Move IndexFindersPart from AbstractList to ArrayList
Since Set doesn't care about collection indexes, we move
IndexFindersPart from AbstractList to ArrayList.
